Audio player extracted to own module, RenderSettings replaced with DisplayOptions
fov is now a slider in options

+++ src/lib-game/src/builder/SceneBuilder.cpp
@@ -98,7 +98,10 @@ static void createCameraAnchorsAtPlaceOfFirstSpawn(
             for (std::size_t idx : std::views::iota(0u, maxPlayerCount))
             {
                 result.emplaceBack(createEntity(
-                    EntityType::CameraAnchor, sf::Vector2f(thing.x, thing.y)));
+                    EntityType::CameraAnchor,
+                    sf::Vector2f(
+                        static_cast<float>(thing.x),
+                        static_cast<float>(thing.y))));
             }
             return;
         }
